+++ docs/CHANGELOG.rst
@@ -5,6 +5,17 @@ All notable changes to the Vector SIL Kit project shall be documented in this fi
 
 The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <http://keepachangelog.com/en/1.0.0/>`_.
 
+
+[4.0.23] - UNRELEASED
+---------------------
+
+Fixed
+~~~~~
+
+- Ethernet (Trivial Sim.): The self-delivery with ``TransmitDirection::TX`` is now triggered only after
+  the frame has been sent with ``TransmitDirection::RX``.
+
+
 [4.0.22] - 2023-04-05
 ---------------------
 
@@ -12,6 +23,7 @@ Fixed
 ~~~~~
 
 - Resolved issue that lead to wrong label matching behavior under certain circumstances.
+
 
 [4.0.21] - 2023-04-03
 ---------------------